requirements-dev.txt
-r requirements.txt
pytest==3.2.3
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.6.1
responses==0.8.1

# Secondary
sqlparse==0.2.4
coverage==4.4.2
py==1.4.34
cookies==2.2.1
